Move joypad into controller mod
