+++ nes/src/lib.rs
@@ -20,8 +20,8 @@ pub mod bus;
 pub mod cartridge;
 pub mod clock;
 pub mod console;
+pub mod controller;
 pub mod cpu;
 pub mod io;
-pub mod joypad;
 pub mod memory;
 pub mod ppu;
